FIX l10n_it_fatturapa_in use fiscal position from partner FIX l10n_it_fatturapa_in: delete einvoice.line when invoice is deleted REF l10n_it_fatturapa_in tests with FatturapaCommon
REF l10n_it_fatturapa_in extracting _compute_e_invoice_validation_error parts for override
