+++ l10n_it_fatturapa_in/tests/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
+from . import fatturapa_common
 from . import test_import_fatturapa_xml
